pyproject.toml
[tool.poetry]
name = "pysnyk"
version = "0.3.1"
description = "A Python client for the Snyk API"
authors = [
  "Gareth Rushgrove <garethr@snyk.io>",
]
license = "MIT"
packages = [
  { include = "snyk" },
]
readme = "README.md"
repository = "https://github.com/snyk-labs/pysnyk"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.22"
mashumaro = "^1.7"
poetry-version = "^0.1.3"
"delegator.py" = "^0.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.0"
pytest-black = "^0.3.7"
pytest-cov = "^2.7"
pytest-mypy = "^0.4.0"
requests-mock = "^1.7"
xlsxwriter = "^1.1.8"
pytest-isort = "^0.3.1"
pytest-mock = "^1.13.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

